Add gnn embeddings from pretraing models (dgl) (#33)
* Compute drug embeddings with pretrained gnns

The drug embeddings are computed in the notebook. A path file is added for better integration, pretrained checkpoints are ignored via .gitignore

* Clean notebook

Remove all unecessary parts from notebook

* Move paths file and notebook

This is to be consistent with the file structure used for the grover embedding.

* Adjust .gitignore

Now all file the respective data folders in the embeddings directory are ignored by git. That means we do not upload any embeddings.parquet file

* Update embedding notebook

This notebook was used to create the embeddings with are on the server. It works out of the box when using the lincs_trapnell.smiles csv file.

* Update paths file

The EMBEDDING_DIR is now the shared one on the cluster.

* Add pretrained GNNs

MPNN, weave, AttentiveFP, and GCN pretrained on  PCBA are added and can be used in CPA.

* Data dir should be a PATH

Co-authored-by: Simon Boehm <simon@siboehm.com>