Fix DE genes for sciplex dataset dependent on available genes (#74)
* Include all observation from the sciplex data

The first notebook simply does the gene matching with the lincs data, ignoring the subsetting from before. The second notebook is anupdated version of the addition of SMILES strings to the `.obs`  dataframe. As a result all `.h5ad` files in the `PROJECT_DIR/'datasets'` folder are updated and ready to be used in our model sweeps.

* Fix JQ1

* Remove redundant code

The cell which renamed the "JQ1" a second time was removed.

* Update relevant sciplex notebooks

The DEGs are compute in the gene matching notebook. The second notebook was run to add the SMILES strings. Addtionally, a random split is added.

* Make key for DE gene sets configurable

The key was hardcoded to `rank_genes_groups_cov`. This is bad as we work with different gene sets and have to be flexible wrt the DEGs.

* Update the manual sweep

The manual sweep is updated to check the new Dataset structure. When the `degs_key` is updated in the SEML config file, everythin works as expected.

* Revert gitignore

Co-authored-by: Simon Boehm <simon@siboehm.com>