Sweep `sciplex_hparam` (#87)
* Add yaml for `sciplex_hparam`

* Decrease number of experiments and fix typo

* Fix typo in config

The typo led to FAILED runs for all grover_base models. I changed the respective property `model.hparams.dosers_depth` also in the mongoDB and restarted the runs.

We wil have to reset the non pre-trained experiments as they have `model.hparams.dosers_depth=3`.

* Add analysis notebook for `sciplex_hparam` sweep

Results are preliminary but they suggest that training on sciplex with the chosen split is too easy. We might have to adjust the split on which we want to evaluate.

* Calc best models separately for pretraining.

* Fix config for seq2seq

* Add drug analysis for sciplex dataset

Based on this analysis an ood split is created which we can use for fine-tuning.

* Add readme for experiment and global experiment setup

* Add config for updated adversary sweep

Only sweep over rdkit and grover, use LINCS hparams and new ood split

* Update analyssis, add plot that shows training performance

* Config: Decrease number of sweeps, change dataset

* Fix config error: split_key

* Fix typo for saving adata object

* Fix split: add controls obs to test set

Split `split_ood_finetuning` had no controls, led to nan values during eval

* Add anlaysis notebook

* Add small sciplex dataset in notebook

Co-authored-by: Simon Boehm <simon@siboehm.com>